+++ src/main/java/com/springboot/moa/post/model/PostDetailsRes.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @AllArgsConstructor
 
 public class PostDetailsRes {
-    int postDetailId;
+    long postDetailId;
 }
